new updates

+++ models/llike.go
@@ -2,7 +2,7 @@ package models
 
 type Like struct {
 	ID    uint `json:"id" gorm:"primary_key"`
-	Stars uint `json:"stars"`
+	Stars uint8 `json:"stars"`
 
 	UserID uint `json:"user_id"`
 	BookID uint `json:"book_id"`
